refactor(destinations): isolate map and views
